My Modifications
This commit and branch contains modifications for OpenMVG's global & incremental SFM (software-only) pipelines in order to keep track of progress. The total bundle adjustment iterations has also been modified from 500 to 100 in order to shorten analysis times. Accuracy of reconstruction for this fork is not guaranteed to be as good as the main. Please use that instead, as modifying the total BA iterations are purely for experimental purposes only.